refactor(resources): change id and price type of `products` table in schema.sql
- id: Long, id used as Long in ProductAPI
- price: Decimal (0.00) format, fixed point
- url: has size limitation — ideal length still needs to be determined

+++ src/main/resources/schema.sql
@@ -1,7 +1,7 @@
 CREATE TABLE products
 (
-    id        BIGINT PRIMARY KEY AUTO_INCREMENT,
+    id        LONG PRIMARY KEY AUTO_INCREMENT,
     name      VARCHAR(255),
-    price     DOUBLE,
+    price     DECIMAL(10, 2),
     image_url VARCHAR(255)
 );
